requirements.txt
# FastApi
fastapi
pydantic
uvicorn
uvicorn-worker
uvicorn[standard]
gunicorn

# Security
pyjwt
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1

# Database
sqlalchemy
sqlalchemy[asyncio]
asyncpg
psycopg2
alembic

# Other
python-dotenv

#Formatters
black